+++ AuthPermissions/IClaimsCalculator.cs
@@ -16,6 +16,7 @@ namespace AuthPermissions
         /// This will return the AuthP claims to be added to the Cookie or JWT token
         /// </summary>
         /// <param name="userId"></param>
+        /// That's because the JWT Token uses a claim of type "nameid" to hold the ASP.NET Core user's ID</param>
         /// <returns></returns>
         Task<List<Claim>> GetClaimsForAuthUserAsync(string userId);
     }