+++ Example2.WebApiWithToken.IndividualAccounts/Controllers/WeatherForecastController.cs
@@ -3,10 +3,8 @@ using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Threading.Tasks;
 using AuthPermissions.AspNetCore;
 using Example2.WebApiWithToken.IndividualAccounts.PermissionsCode;
-using Microsoft.AspNetCore.Authorization;
 
 namespace Example2.WebApiWithToken.IndividualAccounts.Controllers
 {