+++ Test/UnitTests/TestAuthPermissions/TestClaimsCalculatorNoTenant.cs
@@ -4,6 +4,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AuthPermissions;
 using AuthPermissions.BaseCode;
@@ -142,6 +143,5 @@ namespace Test.UnitTests.TestAuthPermissions
             //VERIFY
             claims.Count.ShouldEqual(0);
         }
-
     }
 }